化学品桥接并入 AppliedMekanisticsCompat，判定改用 ModCheckUtils
按评审意见改用仓库已有设施，删掉新引入的两个类：

- MekanismChemicalCompat 的 toGenericStack 并入 AppliedMekanisticsCompat，
  后者本就直接引用 appmek 的 MekanismKey，化学品转换放这里更合适
- MekanismChemicalGate 换成 ModCheckUtils.isAppMekLoading()，
  同时补上 MODID_MEKANISM / MODID_APPMEK 两个常量

「判定与转换分家」这一点仍然成立：判定若与转换写在同一个类里，光是调用判定
就会加载那个类，从而要求 mekanism / appmek 的类可解析，只装 Mekanism 不装
Applied Mekanistics 的环境会直接崩。ModCheckUtils 只读模组列表，满足这一点。

行为不变。
